+++ NEWS.md
@@ -8,8 +8,9 @@
   * debug assign_clust line 15
   * categorization of centers by coverage in leafepp(t = "proy")
   * add assign_nn() function
-  * add addSearchOSM() in 
+  * add addSearchOSM() in leafepp
   * add labels_es dataset to translate leafepp maps 
+  * adaptation to osrm update
 
 ## EPP v0.3.2 (2021-08-03)